Move serialization function for `shared` types to database_utils.hpp.
